refactor: move ssh engine, config, and core events to omnyssh-core
