+++ crates/omnyssh/src/ui/detail_view.rs
@@ -13,10 +13,10 @@ use ratatui::{
 };
 
 use crate::app::{AppAction, AppState, SnippetPopup, ViewState};
-use crate::event::{DetectedService, Metrics, ServiceKind};
-use crate::ssh::client::ConnectionStatus;
 use crate::ui::theme::threshold_color;
 use crate::ui::theme::Theme;
+use omnyssh_core::event::{DetectedService, Metrics, ServiceKind};
+use omnyssh_core::ssh::client::ConnectionStatus;
 
 // ---------------------------------------------------------------------------
 // Render
@@ -142,7 +142,7 @@ pub fn render(frame: &mut Frame, area: Rect, state: &AppState, view: &ViewState)
 fn render_header(
     frame: &mut Frame,
     area: Rect,
-    host: &crate::ssh::client::Host,
+    host: &omnyssh_core::ssh::client::Host,
     metrics: Option<&Metrics>,
     status: Option<&ConnectionStatus>,
     theme: &Theme,
@@ -438,7 +438,7 @@ fn service_hotkey(kind: &ServiceKind) -> String {
 
 /// Get status display for service (e.g., "8 running, 1 stopped", "repl lag: 2.3s").
 fn service_status_display(service: &DetectedService) -> String {
-    use crate::event::MetricValue;
+    use omnyssh_core::event::MetricValue;
 
     match service.kind {
         ServiceKind::Docker => {
